Swap aliases removed in the latest numpy release.
Summary:
Numpy removed support for np.object, np.bool, np.float, np.complex,
np.str, and np.int with 1.24.0 release. Seems like upstream tensorflow
just replaced these with an underscored version of these types, doing
the same here.

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, alfiee

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, alfiee

Maniphest Tasks: T72490

Differential Revision: https://phabricator.sourcevertex.net/D79431

+++ tensorflow/python/ipu/tests/ipu_strategy_test.py
@@ -544,7 +544,7 @@ class IPUStrategyV1Test(test_util.TensorFlowTestCase, parameterized.TestCase):
 
     @def_function.function
     def training_step(num_iteration, iterator):
-      x = constant_op.constant(0.0, np.float)
+      x = constant_op.constant(0.0, np.float32)
       for _ in math_ops.range(num_iteration):
         j = next(iterator)[0]
         outfeed_queue.enqueue(j[0])
